--- src/test/java/by/itacademy/sergeitsarik/SergeiTsarikTest.java
@@ -1,4 +1,0 @@
-package by.itacademy.sergeitsarik;
-
-public class SergeiTsarikTest {
-}
