+++ src/test/java/by/onliner/itacademy/artsiom/SiteTest.java
@@ -1,4 +1,4 @@
-package by.itacademy.artsiom;
+package by.onliner.itacademy.artsiom;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
