+++ src/test/java/by/onliner/itacademy/bazhenava/SiteTest.java
@@ -1,4 +1,4 @@
-package by.itacademy.pavelyatsevich;
+package by.onliner.itacademy.bazhenava;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
@@ -10,28 +10,30 @@ import org.openqa.selenium.WebElement;
 import org.openqa.selenium.chrome.ChromeDriver;
 
 public class SiteTest {
-    WebDriver driver;
+    static WebDriver driver;
 
     @BeforeEach
     public void setupDriver() {
-        String webAddress = "https://taplink.cc/sublimasha";
+        String webAddress = "https://psychologists.taplink.ws";
         driver = new ChromeDriver();
         driver.get(webAddress);
     }
 
     @Test
-    public void openSiteTest() {
-        String fieldSiteFooterTextXPath = "//span[contains(text(), 'Taplink.ru']";
-        By fieldSiteFooterTextBy = By.xpath(fieldSiteFooterTextXPath);
-        WebElement fieldSiteFooterTextWebElement = driver.findElement(fieldSiteFooterTextBy);
-        String actualTextMessage = fieldSiteFooterTextWebElement.getText();
+    public void openAndDownloadSite() {
+        String openAndDownloadTextXPath = "//span[contains(text(), 'Taplink.ru')]";
+        By openAndDownloadTextBy = By.xpath(openAndDownloadTextXPath);
+        WebElement openAndDownloadWebElement = driver.findElement(openAndDownloadTextBy);
+
+        String actualTextMessage = openAndDownloadWebElement.getText();
         String expectedTextMessage = "Taplink.ru";
 
         Assertions.assertEquals(expectedTextMessage, actualTextMessage);
     }
 
     @AfterEach
-    public void teardown() {
+    public void teardown() throws InterruptedException {
+        Thread.sleep(3000);
         driver.quit();
     }
 }
