+++ src/test/java/by/onliner/itacademy/olga/SiteTest.java
@@ -1,4 +1,4 @@
-package by.itacademy.olga;
+package by.onliner.itacademy.olga;
 
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.Assertions;
