+++ src/test/java/by/onliner/itacademy/sergeitsarik/SergeiTsarikTest.java
@@ -1,0 +1,4 @@
+package by.onliner.itacademy.sergeitsarik;
+
+public class SergeiTsarikTest {
+}
